Ui :D
Ui stating to work :+1:

+++ Assets/Script/Data/SerializableCasses.cs
@@ -106,7 +106,7 @@ public class GameResources {
     /// <summary>
     /// Ammout of food.
     /// </summary>
-    public int Food;
+    public float Food;
 
     public void UpdateResources(int w,int s, int i)
     {
